% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_coicop}
\alias{add_coicop}
\title{Details: main function to add coicop categories in the Spanish Household Budget Survey (HBS)}
\usage{
add_coicop(year)
}
\arguments{
\item{year}{year of the HBS you want to use}
}
\description{
Details: main function to add coicop categories in the Spanish Household Budget Survey (HBS)
}
