% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_wh_is}
\alias{adjust_wh_is}
\title{Details: main function to create a graph to summarize the distributional impact based in one or more socioeconomic or demographic variable (one plot per variable)}
\usage{
adjust_wh_is(data, var_w, var_h)
}
\arguments{
\item{data}{description}

\item{var_w}{description}

\item{var_h}{description}
}
\description{
Details: main function to create a graph to summarize the distributional impact based in one or more socioeconomic or demographic variable (one plot per variable)
}
