% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_coicop}
\alias{all_coicop}
\title{List of all COICOP codes available in MEDUSA}
\format{
.csv
}
\source{
medusa
}
\usage{
all_coicop
}
\description{
List of all COICOP codes available in MEDUSA
}
\keyword{datasets}
