% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{av_country}
\alias{av_country}
\title{File containing available countries and country codes}
\format{
.csv
}
\source{
medusa
}
\usage{
av_country
}
\description{
File containing available countries and country codes
}
\keyword{datasets}
