% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{available_var_eu}
\alias{available_var_eu}
\title{available_var_eu}
\usage{
available_var_eu()
}
\value{
variables for which the calculation of basic distributional impacts is available
}
\description{
Function that returns the variables for which the calculation of
basic distributional impacts is available for EU countries.
}
