% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_var_impact}
\alias{available_var_impact}
\title{available_var_impact}
\usage{
available_var_impact()
}
\value{
variables for which the calculation of basic distributional impacts is available
}
\description{
Function that returns the variables for which the calculation of
basic distributional impacts is available.
}
