% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_var_intersec}
\alias{available_var_intersec}
\title{available_var_intersec}
\usage{
available_var_intersec()
}
\value{
set of variables for which the calculation of intersectional
distributional impacts is available
}
\description{
Function that returns the set of variables for which the calculation
of intersectional distributional impacts is available.
}
