% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{available_var_intersec_eu}
\alias{available_var_intersec_eu}
\title{available_var_intersec_eu}
\usage{
available_var_intersec_eu()
}
\value{
set of variables for which the calculation of intersectional
distributional impacts is available for EU countries
}
\description{
Function that returns the set of variables for which the calculation
of intersectional distributional impacts is available for EU countries.
}
