% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basic_graph}
\alias{basic_graph}
\title{Details: main function to create a graph to summarize the distributional impact based in one or more socioeconomic or demographic variable (one plot per variable)}
\usage{
basic_graph(data, var = categories$categories)
}
\arguments{
\item{data}{input data calculate}

\item{var}{socioeconomic or demographic variable/s for which figures are generated. By default: all variables in categories.}
}
\description{
Details: main function to create a graph to summarize the distributional impact based in one or more socioeconomic or demographic variable (one plot per variable)
}
