% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{basic_graph_eu}
\alias{basic_graph_eu}
\title{basic_graph_eu}
\usage{
basic_graph_eu(data, var = graph_labels_eu$VARIABLE)
}
\arguments{
\item{data}{a dataset with the input data needed to generate a basic graph.}

\item{var}{variable(s) according to which you want to generate the graph. If
graph_labels_eu$VARIABLE (by default) creates a graph with the distributional
impacts for each of the variables specified in the package. If not, you can
indicate a variable or a vector of variables to crate the graph.If you want to
see the variables for which the function is available run `available_var_impact_eu()`.}
}
\value{
a graph per selected variable/s summarizing distributional impacts.
}
\description{
Function to create a basic graph to summarize the distributional impact in the EU
based in one or more socioeconomic or demographic variable (one plot per variable).
}
