% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Base.R
\name{basis}
\alias{basis}
\title{Basic function of liner mixed model:
 *Construct fixed/random components and cell size;
 *Perform batch effect correction;
 *Remove outlier cells.}
\usage{
basis(
  exprsData,
  MetaData,
  bulk,
  ct.cell.size,
  data_type,
  Filter,
  BatchCorrect,
  SF,
  gene
)
}
\arguments{
\item{exprsData:}{matrix of the single-cell data (gene x cell);}

\item{MetaData:}{data frame of the summary information of bulk RNA-seq samples. Include sampleID, cellTyp;}

\item{bulk:}{matrix of the bulk RNA-Seq data (gene x cell);}

\item{ct.cell.size:}{vector of cell sizes with labeled cell type names. Default as NULL. If NULL, then estimate cell size from data;}

\item{data_type:}{character, type of the inputed single-cell reference data (count/tpm/rpkm);}

\item{Filter:}{bool, whether to remove outlier cells in single-cell reference data or not;}

\item{BatchCorrect:}{bool, whether to remove the batch effect between bulk data and single-cell reference data or not;}

\item{SF:}{numeric, scaling factor;}

\item{gene:}{vector of signature genes.}
}
\value{
a list with elements:
  *base: matrix of fixed components (gene x cell type);
  *data_cellType:  list matrix of single-cell data (cell type x gene x cell);
  *bulk: matrix of bulk RNA-Seq data  (gene x sample);
  *cellSize: vector of cell sizes with labeled cell type names.
}
\description{
Basic function of liner mixed model:
 *Construct fixed/random components and cell size;
 *Perform batch effect correction;
 *Remove outlier cells.
}
