% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Base.R
\name{BatchCorrection}
\alias{BatchCorrection}
\title{Function for correcting batch effect between bulk RNA-Seq samples and reference single-cell data.}
\usage{
BatchCorrection(exprsData, MetaData, bulk, basis, SF, gene)
}
\arguments{
\item{exprsData:}{matrix of the single-cell data (gene x cell);}

\item{MetaData:}{data frame of the summary information of bulk RNA-seq samples. Include sampleID, cellTyp;}

\item{bulk:}{matrix of the bulk RNA-Seq data (gene x cell);}

\item{basis:}{matrix of fixed components (gene x cell type);}

\item{SF:}{numeric, scaling factor;}

\item{gene:}{vector of signature genes.}
}
\value{
matrix of the corrected bulk RNA-Seq samples (gene x sample).
}
\description{
Function for correcting batch effect between bulk RNA-Seq samples and reference single-cell data.
}
