% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dataset.R
\name{build_dataset}
\alias{build_dataset}
\title{Runs the entire data cleaning procedure for a given year}
\usage{
build_dataset(year, set_root_dir, save_output_to)
}
\arguments{
\item{year}{The SIOP year to filter on}

\item{set_root_dir}{point the function to the folder that contains each year of SIOP data to clean}

\item{save_output_to}{set the CSV save out location}
}
\value{
CSV file
}
\description{
Runs the entire data cleaning procedure for a given year
}
