% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_di}
\alias{calc_di}
\title{Details: main function to calculate distributional impacts for different price shocks}
\usage{
calc_di(input, year, save = T, mapping, elevate = F)
}
\arguments{
\item{input}{from where to load input data options = c(manual, gcam)}

\item{year}{Description}

\item{save}{param2 Description}

\item{mapping}{Description}

\item{elevate}{Description}
}
\description{
Details: main function to calculate distributional impacts for different price shocks
}
