% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_ep}
\alias{calc_ep}
\title{calc_ep}
\usage{
calc_ep(year, index = "all")
}
\arguments{
\item{year}{year/s for energy poverty indices calculation}

\item{index}{energy poverty index or indices you want to calculate. Possible
options: 10%, 2M, LIHC, HEP, HEP_LI. If "all" (by default) calculates
all the indices for the selected year/s.}
}
\value{
a dataframe with the selected energy poverty indices
}
\description{
Function to calculate energy poverty indices
}
