% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{calc_ep_eu}
\alias{calc_ep_eu}
\title{calc_ep_eu}
\usage{
calc_ep_eu(data, index = "all")
}
\arguments{
\item{data}{dataset with the EU HBS data (output from hbs_eu, with COICOP columns
renamed via rename_coicop). Must contain columns: CP045, CP00, CP0411, CP0421,
eq_size, weight, country.}

\item{index}{energy poverty index or indices to be calculated. Possible
options: "10%", "2M", "LIHC", "HEP", "HEP_LI". If "all" (by default) calculates
all the indices for each country.}
}
\value{
a dataframe with the selected energy poverty indices per country.
}
\description{
Function to calculate energy poverty indices for EU countries.
Thresholds are calculated per member state.
}
