% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_tp}
\alias{calc_tp}
\title{calc_tp}
\usage{
calc_tp(year, index = "all")
}
\arguments{
\item{year}{year/s for transport poverty indices calculation}

\item{index}{transport poverty index or indices you want to calculate. Possible
options: 10%, 2M, LIHC, VTU. If "all" (by default) calculates all the indices
for the selected year/s.}
}
\value{
a dataframe with the selected transport poverty indices
}
\description{
Function to calculate transport poverty indices
}
