% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{calc_tp_eu}
\alias{calc_tp_eu}
\title{calc_tp_eu}
\usage{
calc_tp_eu(data, index = "all")
}
\arguments{
\item{data}{dataset with the EU HBS data (output from hbs_eu, with COICOP columns
renamed via rename_coicop). Must contain columns: CP072, CP073, CP0731, CP0732,
CP0733, CP0734, CP0735, CP0411, CP0421, CP00, eq_size, weight, country.}

\item{index}{transport poverty index or indices to be calculated. Possible
options: "10%", "2M", "LIHC", "VTU". If "all" (by default) calculates all the
indices for each country.}
}
\value{
a dataframe with the selected transport poverty indices per country.
}
\description{
Function to calculate transport poverty indices for EU countries.
Thresholds are calculated per member state.
}
