% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{categories}
\alias{categories}
\title{List containing the socioeconomic and demographic categories for the calculation of distributional impacts}
\format{
.csv
}
\source{
medusa
}
\usage{
categories
}
\description{
List containing the socioeconomic and demographic categories for the calculation of distributional impacts
}
\keyword{datasets}
