% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Base.R
\name{cellFilter}
\alias{cellFilter}
\title{Function for remove outlier cells.}
\usage{
cellFilter(exprsData)
}
\arguments{
\item{exprsData:}{matrix of the single-cell data (gene x cell).}
}
\value{
matrix of the filtered single-cell data (gene x cell).
}
\description{
Function for remove outlier cells.
}
