% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{check_var_eu}
\alias{check_var_eu}
\title{check_var_eu}
\usage{
check_var_eu(var)
}
\arguments{
\item{var}{var introduced by the user}
}
\value{
stop if var is wrongly introduced
}
\description{
Check if the parameter var is valid for EU distributional impact calculation
}
