% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_var_impact}
\alias{check_var_impact}
\title{check_var_impact}
\usage{
check_var_impact(var_impact)
}
\arguments{
\item{var_impact}{var_impact introduce by the user}
}
\value{
stop if year is wrongly introduced
}
\description{
Check if the parameter var_impact is valid
}
