% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_year}
\alias{check_year}
\title{check_year}
\usage{
check_year(year)
}
\arguments{
\item{year}{year introduce by the user}
}
\value{
stop if year is wrongly introduced
}
\description{
Check if the parameter year is valid
}
