% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2006}
\alias{coicop_2006}
\title{List to add coicop categories for 2006}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2006
}
\description{
List to add coicop categories for 2006
}
\keyword{datasets}
