% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2007}
\alias{coicop_2007}
\title{List to add coicop categories for 2007}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2007
}
\description{
List to add coicop categories for 2007
}
\keyword{datasets}
