% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2008}
\alias{coicop_2008}
\title{List to add coicop categories for 2008}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2008
}
\description{
List to add coicop categories for 2008
}
\keyword{datasets}
