% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2009}
\alias{coicop_2009}
\title{List to add coicop categories for 2009}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2009
}
\description{
List to add coicop categories for 2009
}
\keyword{datasets}
