% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2010}
\alias{coicop_2010}
\title{List to add coicop categories for 2010}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2010
}
\description{
List to add coicop categories for 2010
}
\keyword{datasets}
