% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2011}
\alias{coicop_2011}
\title{List to add coicop categories for 2011}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2011
}
\description{
List to add coicop categories for 2011
}
\keyword{datasets}
