% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2012}
\alias{coicop_2012}
\title{List to add coicop categories for 2012}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2012
}
\description{
List to add coicop categories for 2012
}
\keyword{datasets}
