% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2013}
\alias{coicop_2013}
\title{List to add coicop categories for 2013}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2013
}
\description{
List to add coicop categories for 2013
}
\keyword{datasets}
