% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2014}
\alias{coicop_2014}
\title{List to add coicop categories for 2014}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2014
}
\description{
List to add coicop categories for 2014
}
\keyword{datasets}
