% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2015}
\alias{coicop_2015}
\title{List to add coicop categories for 2015}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2015
}
\description{
List to add coicop categories for 2015
}
\keyword{datasets}
