% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2016}
\alias{coicop_2016}
\title{List to add coicop categories for 2016}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2016
}
\description{
List to add coicop categories for 2016
}
\keyword{datasets}
