% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2017}
\alias{coicop_2017}
\title{List to add coicop categories for 2017}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2017
}
\description{
List to add coicop categories for 2017
}
\keyword{datasets}
