% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2018}
\alias{coicop_2018}
\title{List to add coicop categories for 2018}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2018
}
\description{
List to add coicop categories for 2018
}
\keyword{datasets}
