% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2019}
\alias{coicop_2019}
\title{List to add coicop categories for 2019}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2019
}
\description{
List to add coicop categories for 2019
}
\keyword{datasets}
