% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2020}
\alias{coicop_2020}
\title{List to add coicop categories for 2020}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2020
}
\description{
List to add coicop categories for 2020
}
\keyword{datasets}
