% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_2021}
\alias{coicop_2021}
\title{List to add coicop categories for 2021}
\format{
.csv
}
\source{
medusa
}
\usage{
coicop_2021
}
\description{
List to add coicop categories for 2021
}
\keyword{datasets}
