% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{coicop_mapping}
\alias{coicop_mapping}
\title{coicop_mapping}
\usage{
coicop_mapping(data, mapping_file = mapping_coicop)
}
\arguments{
\item{data}{dataframe with the hbs data to rename}

\item{mapping_file}{mapping file to be used (included in medusa)}
}
\value{
data file with the renamed expenditure categories
}
\description{
Function to map coicop expenditure categories
}
