% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeAlleles.R
\name{complementAllele}
\alias{complementAllele}
\title{Get Complement of a DNA Allele}
\usage{
complementAllele(allele)
}
\arguments{
\item{allele}{Character string representing an allele (A, T, G, or C).}
}
\value{
Character string with the complement allele.
}
\description{
Returns the Watson-Crick complement: A<->T, G<->C.
}
\examples{
complementAllele("A")  # returns "T"
complementAllele("G")  # returns "C"

}
