% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{computeApproxFStatistic}
\alias{computeApproxFStatistic}
\title{Compute Approximate F-statistic from GWAS Summary Statistics}
\usage{
computeApproxFStatistic(betaZX, seZX)
}
\arguments{
\item{betaZX}{Numeric vector of SNP-exposure effect estimates.}

\item{seZX}{Numeric vector of standard errors for SNP-exposure effects.}
}
\value{
Numeric vector of approximate F-statistics.
}
\description{
Computes the approximation F = (beta_ZX / se_ZX)^2 for each SNP.
  This is the standard approximation used when individual-level data is not
  available.
}
\examples{
computeApproxFStatistic(c(0.5, 0.3), c(0.05, 0.1))

}
