% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convertGenotypeString}
\alias{convertGenotypeString}
\title{Convert VCF-Style Genotype Strings to Integer Allele Dosage}
\usage{
convertGenotypeString(genotypeRaw)
}
\arguments{
\item{genotypeRaw}{Character vector of raw genotype strings.}
}
\value{
Integer vector of allele dosage values (0, 1, or 2). Unrecognized
  values are set to NA.
}
\description{
Converts genotype strings from the OMOP Genomic Extension's
  VARIANT_OCCURRENCE table to integer allele dosage values (0, 1, 2).
  Handles VCF-style genotypes ("0/0", "0/1", "1/1", and phased equivalents
  "0|0", "0|1", "1|1") as well as plain integer strings ("0", "1", "2").
  Unrecognized values are converted to NA with a warning.
}
\examples{
convertGenotypeString(c("0/0", "0/1", "1/1", "0|1", "2"))

}
