% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{country_code}
\alias{country_code}
\title{country_code}
\usage{
country_code()
}
\value{
available countries of the package and their codes
}
\description{
Function that returns the country codes for EU27
}
