% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Estimate.R
\name{CTdcv}
\alias{CTdcv}
\title{Cell-type level deconvolution}
\usage{
CTdcv(
  bulk,
  sce,
  gene = NULL,
  data_type,
  select.ct = NULL,
  RanSplit = NULL,
  ct.cell.size = NULL,
  BatchCorrect = F,
  Filter = T,
  SF = 1000,
  ncpu = NULL,
  iter_max = 1000
)
}
\arguments{
\item{bulk}{A matrix containing bulk RNA-Seq data. Each row corresponds to a certain gene and each column to a certain sample.}

\item{sce}{A 'Seurat' object containing the single-cell RNA-Seq data. Meta data of the 'Seurat' object must includes 'cellType' and 'sampleID'.}

\item{gene}{A character vector of the gene names to use as signature for the deconvolution. We summarized signature genes for the 64 human cell-types. please see 'sg'.}

\item{data_type}{A character of the type of the single-cell RNA-Seq data, including 'count', 'tpm', 'rpkm','fpkm', and 'cpm'.}

\item{select.ct}{A character vector of the names of the target cell-types. The default value is NULL. With default value, all cell-types in the single-cell data will be used.}

\item{RanSplit}{A character (or character vector) to split the random components. The default value is NULL. With default value, all cells, excepting those in target the cell-type, will be fitted as one random component.}

\item{ct.cell.size}{A character vector of the cell-size (total mRNA amount) of the selected cell-types. The default value is NULL.}

\item{BatchCorrect}{A Boolean variable to determine whether to run 'ComBat' to correct batch effects between single-cell RNA-Seq and bulk RNA-Seq or not. The default value is FALSE.}

\item{Filter}{A Boolean variable to determine whether to filter the outlier in single-cell data or not. The default value is FALSE.}

\item{SF}{Scaling factor. The default value is 1e+3.}

\item{ncpu}{The number of CPU cores to be used.}

\item{iter_max}{The maximum iterations of REML.}
}
\value{
A list with elements:
  *ct.pro: matrix of cell-type proportions estimated by mixed linear model (sample x cell-type);
  *ct.pro.p: matrix of p value (χ²(df=1)) for the cell-type proportions estimated by the mixed linear model (sample x cell-type);
  *cellSize: vector of cell sizes with labeled cell-type names.
}
\description{
Cell-type level deconvolution
}
\examples{
library(MLM)
ct.es = CTdcv(bulk = example.bulk,sce = example.sce,gene = example.gene,data_type = 'count')

}
