% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{database_hbs}
\alias{database_hbs}
\title{database_hbs}
\usage{
database_hbs(year, country = "all", inputs_path)
}
\arguments{
\item{year}{year of the HBS to process. Available options: 2015, 2020.}

\item{country}{code of the country or countries of the HBS to process. By default, it processes all available data in the working directory To see the available countries and codes, run country_code().}

\item{inputs_path}{Local path to the folder where the HBS's are stored. Not included in the package.}
}
\value{
RData file with the joint database with the HBS microdata.
}
\description{
Function to load each country HBS and create a joint database for household and member files
}
