% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epf_2013_h}
\alias{epf_2013_h}
\title{Household file of the 2013 Household Budget Survey (INE), which includes 8 sections with information on each of the households in the sample: 1) General information, 2) Characteristics of the household, 3) Characteristics of the household reference person, 4) Characteristics of the main dwelling, 5) Other dwellings available to the household, 6) Household consumption expenditures, 7) Regular monthly household income and 8) Number of lunches and dinners during the two-week period.}
\format{
.csv
}
\source{
medusa
}
\usage{
epf_2013_h
}
\description{
Household file of the 2013 Household Budget Survey (INE), which includes 8 sections with information on each of the households in the sample: 1) General information, 2) Characteristics of the household, 3) Characteristics of the household reference person, 4) Characteristics of the main dwelling, 5) Other dwellings available to the household, 6) Household consumption expenditures, 7) Regular monthly household income and 8) Number of lunches and dinners during the two-week period.
}
\keyword{datasets}
