% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epf_2014_g}
\alias{epf_2014_g}
\title{Expenditure file of the 2014 Household Budget Survey (INE), which includes the expenditures at the highest COICOP disaggregation level for each of the households in the sample.}
\format{
.csv
}
\source{
medusa
}
\usage{
epf_2014_g
}
\description{
Expenditure file of the 2014 Household Budget Survey (INE), which includes the expenditures at the highest COICOP disaggregation level for each of the households in the sample.
}
\keyword{datasets}
