% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epf_2021_g}
\alias{epf_2021_g}
\title{descripción de este archivo, copiar del excel}
\format{
.csv
}
\source{
medusa
}
\usage{
epf_2021_g
}
\description{
descripción de este archivo, copiar del excel
}
\keyword{datasets}
