% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Estimate.R
\name{Estimate}
\alias{Estimate}
\title{Bulk RNA-seq deconvolution analysis with lienar mixed model}
\usage{
Estimate(
  bulk,
  sce,
  gene,
  data_type,
  select.ct = NULL,
  RanSplit = NULL,
  ct.cell.size = NULL,
  BatchCorrect = F,
  Filter = T,
  SF = 1000,
  DoParallel = FALSE,
  ncpu = NULL,
  verbose = FALSE,
  iter_max = 1000
)
}
\arguments{
\item{bulk:}{matrix of bulk tissue expression;}

\item{sce:}{object of SingleCellExperiment, single-cell reference data;}

\item{gene:}{vector of signature genes;}

\item{data_type:}{character, type of the inputed single-cell reference data (count/tpm/rpkm);}

\item{select.ct:}{vector of cell types included, default as \code{NULL}. If \code{NULL}, include all cell types in \code{x};}

\item{RanSplit:}{character, label for how to seperate random components. Default as \code{NULL}. If \code{NULL}, all cells will be fitted as one random component;}

\item{ct.cell.size:}{vector of cell sizes with labeled cell type names. Default as NULL. If NULL, then estimate cell size from data;}

\item{BatchCorrect:}{bool, whether to remove the batch effect between bulk data and single-cell reference data or not. Default as FALSE;}

\item{Filter:}{bool, whether to remove outlier cells in single-cell reference data or not. Default as TRUE;}

\item{SF:}{numeric, scaling factor.Default as 1e+3;}

\item{DoParallel:}{bool, whether to perform analysis parallelly or not.Default as FALSE;}

\item{ncpu:}{numeric, the number of cpus used. Default as NULL. If NULL, then maximum detected cpus - 1 will be used ;}

\item{verbose:}{bool, whether to show running information or not during deconvolution analysis. Default as FALSE;}

\item{iter_max:}{numeric, maximum iteration number.Default as 1e+3.}
}
\value{
a list with elements:
  *ct.pro: matrix of cell type proportions estimated by mixed linear model (sample x cell type);
  *ct.pro.p: matrix of p value (χ²(df=1)) for the cell type proportions estimated by mixed linear model (sample x cell type);
  *cellSize: vector of cell sizes with labeled cell type names.
}
\description{
Bulk RNA-seq deconvolution analysis with lienar mixed model
}
\examples{
es = Estimate(bulk = bulk,sce = sce,gene = gene,select.ct = select.ct,data_type = 'count')


}
