% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ex_shocks}
\alias{ex_shocks}
\title{ex_shocks}
\usage{
ex_shocks()
}
\value{
a csv file in which the change in prices must be introduced to apply a price shock
}
\description{
Function to save a csv file in which the change in prices must be
introduced to apply a price shock. Save the file, introduce the price shocks
and load in R.The COICOP variables of the file correspond to the aggregate
variables of the package, if you are not going to aggregate the COICOP
variables you have to replace the column labels by the COICOP variables that
appear in your dataset.
}
