% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{ex_shocks_eu}
\alias{ex_shocks_eu}
\title{ex_shocks_eu}
\usage{
ex_shocks_eu()
}
\value{
a csv file in which the change in prices must be introduced to apply a price shock to EU countries
}
\description{
Function to save a csv file in which the change in prices must be
introduced to apply a price shock for EU countries. Save the file,
introduce the price shocks and load in R.
}
