% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ex_var_intersec}
\alias{ex_var_intersec}
\title{ex_var_intersec}
\usage{
ex_var_intersec()
}
\value{
a csv file with the set of variables for which the calculation of
intersectional distributional impacts is available
}
\description{
Function to save a csv file in which the set of variables must be
introduced for the calculation of intersectional distributional impacts.
Save the file, keep just the conbinations you are interested in and load in R.
}
