% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_eu.R
\name{ex_var_intersec_eu}
\alias{ex_var_intersec_eu}
\title{ex_var_intersec_eu}
\usage{
ex_var_intersec_eu()
}
\value{
a csv file with the set of variables for which the calculation of
intersectional distributional impacts is available for EU countries
}
\description{
Function to save a csv file in which the set of variables must be
introduced for the calculation of intersectional distributional impacts
for EU countries. Save the file, keep just the combinations you are
interested in and load in R.
}
