% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcfhogares95_22}
\alias{gcfhogares95_22}
\title{File containing household expenditure by COICOP category from the national accounts}
\format{
.csv
}
\source{
medusa
}
\usage{
gcfhogares95_22
}
\description{
File containing household expenditure by COICOP category from the national accounts
}
\keyword{datasets}
