% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_attendance.R
\name{get_conf_attendance}
\alias{get_conf_attendance}
\title{Cleans a tibble containing SIOP annual conference participants}
\usage{
get_conf_attendance(df, merge_demo_from, year)
}
\arguments{
\item{df}{a tibble of Annual Conference participants}

\item{merge_demo_from}{The resulting tibble from  make_final_data}

\item{year}{Specify conference year as numeric vector}
}
\value{
a tibble of conference attendees
}
\description{
Cleans a tibble containing SIOP annual conference participants
}
