% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcam_prices.R
\name{get_prices_gcam}
\alias{get_prices_gcam}
\title{get_prices_gcam}
\usage{
get_prices_gcam(
  db_path = NULL,
  query_path = "inst/extdata",
  db_name = NULL,
  prj_name,
  prj = NULL,
  scenarios,
  queries = "queries_GCAM_MEDUSA.xml",
  final_db_year = 2100,
  saveOutput = T,
  base_scen,
  selected_year
)
}
\arguments{
\item{db_path}{Path to the GCAM database}

\item{query_path}{Path to the query file}

\item{db_name}{Name of the GCAM database}

\item{prj_name}{Name of the rgcam project. This can be an existing project, or, if not, this will be the name}

\item{prj}{rgcam loaded project}

\item{scenarios}{Vector names of the GCAM scenarios to be processed}

\item{queries}{Name of the GCAM query file. The file by default includes the queries required to run rfasst}

\item{final_db_year}{Final year in the GCAM database (this allows to process databases with user-defined "stop periods")}

\item{saveOutput}{Writes the files.By default=T}

\item{base_scen}{The base scenario that other scenarios will be compared against}

\item{selected_year}{The year of analysis, when the scenarios wil be compared}
}
\value{
a dataframe with prices by sector and GCAM-Europe region
}
\description{
Extract and format prices from a GCAM-Europe database or project file for MEDUSA
}
\keyword{GCAM,}
\keyword{GCAM-Europe,}
\keyword{prices}
