% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_labels}
\alias{graph_labels}
\title{List containing long names for the socioeconomic and demographic categories to introduce them on the axes of the graphs}
\format{
.csv
}
\source{
medusa
}
\usage{
graph_labels
}
\description{
List containing long names for the socioeconomic and demographic categories to introduce them on the axes of the graphs
}
\keyword{datasets}
