% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_ep2}
\alias{id_ep2}
\title{id_ep2}
\usage{
id_ep2(data)
}
\arguments{
\item{data}{dataset with the data from the HBS.}
}
\value{
a dataset with HBS data where energy poor households are identified.
}
\description{
Function to identify energy poor households before 2016 (included)
}
