% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{id_ep_eu}
\alias{id_ep_eu}
\title{id_ep_eu}
\usage{
id_ep_eu(data)
}
\arguments{
\item{data}{dataset with the EU HBS data (output from hbs_eu, with COICOP columns renamed via rename_coicop).}
}
\value{
a dataset with EU HBS data where energy poor households are identified per country.
}
\description{
Function to identify energy poor households in EU countries.
Thresholds are calculated per member state.
}
