% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_tp}
\alias{id_tp}
\title{id_tp}
\usage{
id_tp(data, year)
}
\arguments{
\item{data}{dataset with the data from the HBS.}

\item{year}{year of the HBS data.}
}
\value{
a dataset with HBS data where transport poor households are identified.
}
\description{
Function to identify transport poor households before 2015 (included)
}
