% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_identify.R
\name{identify_hmdb}
\alias{identify_hmdb}
\title{MZLOG-OBJ PCA}
\usage{
identify_hmdb(mzs, adduct, hmdb_file = "/home/rstudio/hmbd_metabolites.xml")
}
\arguments{
\item{mzs}{\cr
List : List of MZs to Identify}

\item{adduct}{\cr
String: Name of a common adduct
Integer: Atomic mass of an adduct}

\item{hmdb_file}{\cr
String: Location of hmdb.xml download file}
}
\description{
Not really sure
\itemize{
\item Requires: ggplot2, tibble
}
}
