% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impact}
\alias{impact}
\title{Details: main function to apply a specific price shock to the different COICOP categories}
\usage{
impact(data)
}
\arguments{
\item{data}{input data calculate}
}
\description{
Details: main function to apply a specific price shock to the different COICOP categories
}
