% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileIO.R
\name{importSiteProfile}
\alias{importSiteProfile}
\title{Import a Site Profile from CSV}
\usage{
importSiteProfile(profilePath, metadataPath = NULL)
}
\arguments{
\item{profilePath}{Character. Path to the profile CSV file (the file
containing beta and log_likelihood columns).}

\item{metadataPath}{Character. Path to the metadata CSV file. If NULL
(default), the function infers the path by replacing "_profile_" with
"_metadata_" in \code{profilePath}.}
}
\value{
A list with the same structure as \code{\link{fitOutcomeModel}}
  output, suitable for passing to \code{\link{poolLikelihoodProfiles}}.
}
\description{
Reads a site profile from CSV files previously written by
\code{\link{exportSiteProfile}}. Reconstructs the profile list object
that can be passed to \code{\link{poolLikelihoodProfiles}}.
}
\examples{
\dontrun{
profile <- importSiteProfile("medusa_profile_site_A.csv")
}

}
\seealso{
\code{\link{exportSiteProfile}}, \code{\link{poolLikelihoodProfiles}}
}
