% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersectional_graph}
\alias{intersectional_graph}
\title{Details: main function to create a graph to summarize the distributional impact based in the intersection of two socioeconomic or demographic variables (2 variables per plot)}
\usage{
intersectional_graph(data, pairs = is_categories)
}
\arguments{
\item{data}{input data calculate}

\item{var1}{socioeconomic or demographic variable/s................ By default: all variables in categories}

\item{var2}{socioeconomic or demographic variable/s................ By default: all variables in categories}
}
\description{
Details: main function to create a graph to summarize the distributional impact based in the intersection of two socioeconomic or demographic variables (2 variables per plot)
}
