% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{is_categories}
\alias{is_categories}
\title{List containing the socioeconomic and demographic categories for the calculation of intersectional distributional impacts}
\format{
.csv
}
\source{
medusa
}
\usage{
is_categories
}
\description{
List containing the socioeconomic and demographic categories for the calculation of intersectional distributional impacts
}
\keyword{datasets}
