% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{is_categories_eu}
\alias{is_categories_eu}
\title{List containing the socioeconomic and demographic categories for the calculation of intersectional distributional impacts for EU countries}
\format{
.csv
}
\source{
medusa
}
\usage{
is_categories_eu
}
\description{
List containing the socioeconomic and demographic categories for the calculation of intersectional distributional impacts for EU countries
}
\keyword{datasets}
