% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isStrandAmbiguous}
\alias{isStrandAmbiguous}
\title{Check for Strand-Ambiguous SNPs}
\usage{
isStrandAmbiguous(effectAllele, otherAllele)
}
\arguments{
\item{effectAllele}{Character vector of effect alleles.}

\item{otherAllele}{Character vector of other alleles.}
}
\value{
Logical vector indicating which SNPs are strand-ambiguous.
}
\description{
Identifies SNPs with ambiguous strand (A/T or G/C allele pairs).
}
\examples{
isStrandAmbiguous(c("A", "G", "A"), c("T", "C", "C"))

}
