% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lists}
\alias{lists}
\title{List to add coicop categories}
\format{
.csv

.csv
}
\source{
medusa

medusa
}
\usage{
lists

lists
}
\description{
List to add coicop categories

List to apply a specific price shock to the different COICOP categories
}
\keyword{datasets}
