% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_epf}
\alias{load_epf}
\title{Details: main function to load the Spanish epf}
\usage{
load_epf(input, year, save = T, mapping, elevate = F)
}
\arguments{
\item{input}{from where to load input data options = c("manual", "gcam")}

\item{year}{Description}

\item{save}{param2 Description}

\item{mapping}{Description}

\item{elevate}{Description}
}
\description{
Details: main function to load the Spanish epf
}
