% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_hbs}
\alias{load_hbs}
\title{Details: main function to load the Spanish Household Budget Survey (HBS)}
\usage{
load_hbs(year)
}
\arguments{
\item{year}{year of the HBS you want to load}
}
\description{
Details: main function to load the Spanish Household Budget Survey (HBS)
}
