% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demo_data.R
\name{make_demo_data}
\alias{make_demo_data}
\title{A function wrapper that executes all functions related to cleaning the
demographic xlsx dataset.}
\usage{
make_demo_data(df, ...)
}
\arguments{
\item{df}{a tibble of demographic data}

\item{...}{detailed_types for ethnicity breakdown, default is FALSE}
}
\value{
a tibble with all demographic columns cleaned
}
\description{
A function wrapper that executes all functions related to cleaning the
demographic xlsx dataset.
}
\examples{
#make_demo_data(df = demo_raw)
}
