% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dues.R
\name{make_dues}
\alias{make_dues}
\title{Bind multiple rows from dues files}
\usage{
make_dues(..., year)
}
\arguments{
\item{...}{Specify each tibble containing dues data}

\item{year}{Specify the SIOP membership year}
}
\value{
tibble
}
\description{
Takes and merges multiple tibbles into a single tibble in order to
then run make_final_data().
}
