% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_final_data.R
\name{make_final_data}
\alias{make_final_data}
\title{A function wrapper that takes the cleaned dues and demo data and merges it,
and cleans by SID, and finalizes a cleaned dataset ready for further analysis.}
\usage{
make_final_data(demo, dues, staff)
}
\arguments{
\item{demo}{a tibble of demo data. The output from make_demo_data()}

\item{dues}{a tibble of dues data. The output from make_dues_data()}
}
\value{
a tibble of SIOP members for the current year.
}
\description{
A function wrapper that takes the cleaned dues and demo data and merges it,
and cleans by SID, and finalizes a cleaned dataset ready for further analysis.
}
\examples{
#final_data <- make_final_data(demo = demo, dues = dues, staff = staff)
}
