% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANOVA_Pro.R
\name{MANOVA_Pro}
\alias{MANOVA_Pro}
\title{MANOVA_Pro: Combines multiple analysis of variance (MANOVA) with polynomial regression to detect differences in cell-state abundance among groups.}
\usage{
MANOVA_Pro(MeDuSA_obj, degree = NULL, condition = NULL)
}
\arguments{
\item{MeDuSA_obj}{The MeDuSA object.}

\item{degree}{A numeric variable used to specify the polynomial degrees.}

\item{condition}{A character vector containing the biological condition for each bulk sample.}
}
\description{
MANOVA_Pro: Combines multiple analysis of variance (MANOVA) with polynomial regression to detect differences in cell-state abundance among groups.
}
