% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping}
\alias{mapping}
\title{Mapping to rename socioeconomic variables}
\format{
.csv
}
\source{
medusa
}
\usage{
mapping
}
\description{
Mapping to rename socioeconomic variables
}
\keyword{datasets}
