% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_coicop}
\alias{mapping_coicop}
\title{Mapping for expenditure categories (COICOP)}
\format{
.csv
}
\source{
medusa
}
\usage{
mapping_coicop
}
\description{
Mapping for expenditure categories (COICOP)
}
\keyword{datasets}
