% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCTD.R
\name{MCTD}
\alias{MCTD}
\title{MCTD: Mixed model-based Cell Trajectory Deconvolution.}
\usage{
MCTD(
  bulk,
  sce,
  selectCellType,
  ncpu = 1,
  smooth = TRUE,
  gene = NULL,
  nbins = 10,
  resolution = 50,
  knots = 10,
  maxgene = 200,
  mode = "gaussian",
  gcov = NULL,
  Xc = NULL,
  maxiter = 10000
)
}
\arguments{
\item{bulk}{A matrix of bulk RNA-Seq data. Each row corresponds to a specific gene and each column corresponds to a particular sample.}

\item{sce}{A \code{Seurat} object of the single-cell RNA-Seq data (see \code{\link{Seurat}}). Meta data of the 'sce' need include 'cellType' and 'cellTrjaectory'.}

\item{selectCellType}{A character of the names of the target cell-type.}

\item{ncpu}{The number of CPU cores to be used.}

\item{smooth}{A Boolean variable to determine whether to smooth the cell-abundance along the cell-trajectory or not. The default value is \emph{TRUE}. With default, the predicted cell abundance will be smoothed along the cell trajectory via the Loess (see \code{\link{loess}}).}

\item{gene}{A character vector of the genes to use as signatures.The default value is \emph{NULL}. With default, \code{MCTD} will select genes via the generalized additive model (GAM,see \code{\link{mgcv}}).}

\item{nbins}{A numeric variable to determine the number of cell-bins used in the gene selection. The parameter \code{nbins} was used to ensure the selected genes were evenly distributed along the cell trajectory. The default value is \emph{10}.}

\item{resolution}{A numeric variable to determine the resolution of deconvolution. After setting \code{resolution}, cells will be grouped into \emph{"resolution"} cell-bins. The default value is \emph{50}.}

\item{knots}{A numeric variable to specify the numbers of knots used in the GAM. The default value is \emph{10}.}

\item{maxgene}{A numeric variable to specify the maximum numbers of genes to be selected. The default value is \emph{200}.}

\item{mode}{A character variable to specify the distribution and link to use in the gene selection. See \code{\link{family.mgcv}} for a full list of what is available. Note that we suggest to use the \emph{gaussian} or the \emph{nb} (negative binomial distribution)' here.The default mode is the \emph{gaussian}.}

\item{gcov}{A matrix (or vector) of fixed covariates in gene selection,such as donors, genders and extra.The default value is \emph{NULL}.}

\item{Xc}{A matrix (or vector) of fixed covariates in deconvolution.The default value is \emph{NULL}. With default, \code{MCTD} will include expression profiles of other cell-types as the deconvolution covariates.}

\item{maxiter}{The maximum iterations of REML.The default value is \emph{1e+4}.}
}
\value{
\code{MCTD} returns: \itemize{
\item\code{abundance}: A matrix of cell-abundance. Each row corresponds to a median trajectory point of the cell-bins and each column corresponds to a certain sample.
\item\code{gene}: A vector of the signature genes used in deconvolution.
}
}
\description{
This is the main function of the "MCTD" method. \code{MCTD} is used to predict cell-abundance along a predefined cell trajectory in the given bulk RNA-seq data. \code{MCTD} is well suitable for biological scenarios in which the underlying mechanisms are associated with continuous transitions of cellular states.
}
\details{
Unlike most deconvolution methods decomposing cell-abundance in bulk tissues for cell-types,
\code{MCTD} is a fine-resolution deconvolution method that predicts cell abundance along a predefined cell trajectory.
Cell-trajectory specifies each cell as a point in a geometric vector, determining the pattern of a dynamic process experienced by cells.
With inputting cell trajectory and scRNA-seq as references, \code{MCTD} infers the cell-abundance along the cell trajectory within the given bulk RNA-seq data.
\code{MCTD} contains three steps:\itemize{
\item Gene selection: \code{MCTD} selects signature genes to distinguish cells along the cell-trajectory via the GAM.
\item Deconvolution: \code{MCTD} infers the cell-abundance for each cell-bin via the mixed model.
\item Smoothing: \code{MCTD} smooths the cell-abundance along the cell-trajectory via the loess.This step is performed if smooth = TRUE.}
}
\examples{
##Load the test data
data(ref)
data(cellType)
data(cellTrjaectory)
data(bulk)

##Build the 'Seurat' obejct:
sce = CreateSeuratObject(ref)
sce$cellType = cellType
sce$cellTrjaectory = rep(0,ncol(sce))
sce$cellTrjaectory[rownames(Trajectory)]=Trajectory

##Run MCTD (with 6 CPU cores)
CellAbundance = MCTD(bulk=bulk,sce=sce,selectCellType='Epithelium',ncpu=6)$abundance
}
\author{
Liyang Song <songliyang@westlake.edu.cn>
}
