% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDuSA.R
\name{MeDuSA}
\alias{MeDuSA}
\title{MeDuSA: mixed model-based deconvolution of cell-state abundance.}
\usage{
MeDuSA(
  bulk,
  sce,
  select.ct,
  ncpu = 1,
  smooth = TRUE,
  smoothMethod = "loess",
  gene = NULL,
  nbins = 10,
  resolution = 50,
  knots = 10,
  start = c(1e-05, 0.01),
  maxgene = 200,
  family = "gaussian",
  gcov = NULL,
  Xc = NULL,
  maxiter = 10000,
  adj = FALSE,
  CAR = TRUE,
  phi = c(0.2, 0.4, 0.6, 0.9)
)
}
\arguments{
\item{bulk}{A matrix of bulk RNA-seq data. Each row corresponds to a specific gene and each column corresponds to a particular sample.}

\item{sce}{A \code{seurat} object of the  reference scRNA-seq data (see \code{\link{seurat}}). Meta-data of the seurat object (sce@meta.data) need include two columns named as cell_type and cell_trajectory.}

\item{select.ct}{A character variable of the target cell type.}

\item{ncpu}{The number of CPU(s) to be used.}

\item{smooth}{A Boolean variable to determine whether to smooth the predicted cell-state abundance or not. The default value is TRUE.}

\item{smoothMethod}{A character variable to determine the smoothing method. The default method is \code{\link{loess}}.}

\item{gene}{A character vector of signature genes. The default value is NULL. With default, \code{MeDuSA} selects signature genes using the generalized additive model (see \code{\link{mgcv}}).}

\item{nbins}{A numeric variable to determine the number of bins along the cell trajectory, which is used to ensure the selected genes were uniformly scattered along the given trajectory. The default value is 10.}

\item{resolution}{A numeric variable to determine the number of cell bins along the cell trajectory (i.e., the resolution of the deconvolution analysis). The default value is 50.}

\item{knots}{A numeric variable to specify the number of knots. The default value is 10.}

\item{start}{A numeric vector for the initial value of the REML. Default by c(1e-5,1e-2).}

\item{maxgene}{A numeric variable to determine the number of signature genes. The default value is 200.}

\item{family}{A character variable to specify the distribution of GAM. See \code{\link{family.mgcv}} for a full list of what is available.}

\item{gcov}{A matrix (vector) of covariates in the GAM. The default value is NULL.}

\item{Xc}{A matrix (vector) of fixed covariates in the linear mixed model (i.e., covariates for predicting cell-state abundance). The default value is NULL.}

\item{maxiter}{The iteration number of REML. Default by 1e+4.}

\item{adj}{A Boolean variable to determine whether to include covariates when predicting the cell-state abundance.}

\item{CAR}{A Boolean variable to determine whether to model abundance correlations among cells.}

\item{phi}{A numeric vector for searching the optimal cell correlations. The default value is c(0.2,0.4,0.6,0.9)}
}
\value{
\code{MeDuSA} returns: \itemize{
\item\code{abundance}: A numeric matrix of cell-state abundance. Each row corresponds to a cell bin and each column corresponds to a sample.
\item\code{gene}: A character vector of the selected signature genes.
\item\code{PesudoTimeCellbin}: A numeric vector of the pseudo-time for each cell bin.
}
}
\description{
\code{MeDuSA} is a fine-resolution cellular deconvolution method that leverages scRNA-seq data as a reference to estimate cell-state abundance in bulk RNA-seq data.
}
\examples{
##Library the package
library(MeDuSA)

##Load the test data:
data(ref)
data(cellType)
data(Trajectory)
data(bulk)

##Build the 'Seurat' obejct:
sce = CreateSeuratObject(ref)
sce$cell_type = cellType
sce$cell_trajectory = rep(0,ncol(sce))
sce$cell_trajectory[rownames(Trajectory)]=Trajectory

##Run MeDuSA (run with 6 courses):
csab = MeDuSA(bulk=bulk,sce=sce,select.ct='Epithelium',ncpu=6)
}
\author{
Liyang Song <songliyang@westlake.edu.cn>
}
