% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDuSA_VarExplain.R
\name{MeDuSA_VarExplain}
\alias{MeDuSA_VarExplain}
\title{MeDuSA_VarExplain: the function to estimate explained variances of bulk RNA-seq data by the reference scRNA-seq data.}
\usage{
MeDuSA_VarExplain(MeDuSA_obj, maxiter = 1000, ncpu = 4)
}
\arguments{
\item{MeDuSA_obj}{The MeDuSA object.}

\item{maxiter}{The iteration number of REML. Default by 1e+4.}

\item{ncpu}{The number of cpu cores to be used.}
}
\value{
\code{MeDuSA} returns: \itemize{
\item\code{MeDuSA_Object@VarianceExplain$VarExplain}: A numeric vector of the explained variances of bulk RNA-seq data by the reference scRNA-seq data.
\item\code{MeDuSA_Object@VarianceExplain$Chi_VarExplain}: A numeric vector of the corresponding squared-chi-value of the explained variances.
\item\code{MeDuSA_Object@VarianceExplain$Pval_VarExplain}: A numeric vector of the corresponding p-value of the explained variances.
}
}
\description{
MeDuSA_VarExplain: the function to estimate explained variances of bulk RNA-seq data by the reference scRNA-seq data.
}
