% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_mz_post_processing.R
\name{mz_post_process_magic}
\alias{mz_post_process_magic}
\title{MZ-OBJ Process Magic}
\usage{
mz_post_process_magic(input_mz_obj, metadata, plot = TRUE)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{metadata}{\cr
DataFrame : Input MZ-Metadata}

\item{plot}{\cr
Boolean   : to plot or not to plot}
}
\value{
c(mzLong_obj, mzLog_obj)
}
\description{
Do all the recommended post processing
\itemize{
\item Requires: ggplot2, tidyrx
}
}
