% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_mz_quality_check.R
\name{mz_quality_metrics}
\alias{mz_quality_metrics}
\title{MZ-OBJ Quality Metrics}
\usage{
mz_quality_metrics(input_mz_obj, cores = 2)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{cores}{Integer: Can I has multithreading? (Need parallel)

Dependencies : dplyr}
}
\value{
Returns a dataframe of samples by metrics
}
\description{
Get Statistical Mz Data\cr
}
