% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_mz_subtraction.R
\name{mz_subtraction}
\alias{mz_subtraction}
\title{MZ-OBJ Subtraction (usually used for blanks subtraction)}
\usage{
mz_subtraction(
  input_mz_obj,
  samples,
  compare_samples,
  method = mean,
  threshold = 3
)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{samples}{\cr
DataFrame : Metadata-Obj of Samples}

\item{compare_samples}{\cr
DataFrame : Metadata-Obj of Compared samples (blanks)}

\item{threshold}{\cr
Int       : "Delete samples < compare_samples * threshold"}

\item{method:}{\cr
String : method to apply to compare samples before threshold}
}
\value{
Returns an MZ-OBJ
}
\description{
Compares the intensities of two "types" of samples.\cr
Zeros any sample-intensities
that are lower  than "Threshold"*"ComparedIntensities"
}
\examples{

To remove values from "samples" that are lower than "blanks" * "threshold"


blanks_meta <- filter(metadata, sampletype == "blank" &
                                  ionmode == ionmode_val & phase == phase_val &
                                  filtered_out == "no")
samples_meta <- filter(metadata, sampletype == "sample" &
                                  ionmode == ionmode_val  & phase == phase_val &
                                  filtered_out == "no")

mz_subtraction(input_mz_obj, samples_meta, blanks_meta, threshold = "5" )
}
