% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_analysis.R
\name{mzlog_analysis_welch}
\alias{mzlog_analysis_welch}
\title{MZLOG-OBJ T-Test}
\usage{
mzlog_analysis_welch(
  input_mzlog_obj,
  phenotype_a,
  phenotype_b,
  adjust = "fdr",
  cores = 2
)
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}

\item{cores}{Integer: Can I has multithreading? (Need parallel)}

\item{adjust_method}{\cr
String: How to p.adjust?
Boolean: False = no adjustment ; True = "fdr"

Dependencies : pbapply, dplyr}
}
\value{
DataFrame with columns:
\itemize{
\item mz      : Float: MZ
\item p       : Float: p-value
\item p_05    : Boolean: is P < 0.05
\item p_10    : Boolean: is P < 0.10
\item p_15    : Boolean: is P < 0.15
\item adjusted: p.adjust($p, adjust_method) *optional, but default
}
}
\description{
Welch T-test, input should be an log2 of mz_obj
\itemize{
\item Requires: dplyr
'
}
}
