% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_randomforest.R
\name{mzlog_rf_train}
\alias{mzlog_rf_train}
\title{Train you model data within a mzLog_obj}
\usage{
mzlog_rf_train(
  mzlog_obj,
  rfe_obj,
  meta,
  attribute = "phenotype",
  seeds = c(42, 666, 314159, 1.05457, 998001),
  ratio = 0.8,
  cores = 4
)
}
\arguments{
\item{attribute}{\cr
String: which metadata attribute to compare}

\item{ratio}{\cr
Float: split ratio for train}

\item{correlation_data}{\cr
DataFrame : from rf_correlation}

\item{metadata}{\cr
DataFrame: metadata object}

\item{feat_size_seq}{\cr
Sequence to find optimal "number_of_variables"}

\item{seed}{\cr
List: which seeds to use. Also how many runs to do}
}
\value{
caret::rfe object
}
\description{
Train you model data within a mzLog_obj
}
