% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_mzml_extract_centroid.R
\name{mzml_extract_file}
\alias{mzml_extract_file}
\title{Return mzT from a single file}
\usage{
mzml_extract_file(
  file,
  polarity = NULL,
  magic = T,
  massWindow = c(0, Inf),
  cl = NULL
)
}
\arguments{
\item{file}{\cr
String : Filename.mzML}

\item{polarity}{\cr
\link{0|1}: 0=Negative, 1=Positive, NULL=both}

\item{magic}{\cr
Boolean : To run default filtering, and TimeSquashing}

\item{massWindow}{\cr
c(Int,Int): Filter the mass window}

\item{cl}{\cr
parallel::threadCluster (optional)}
}
\description{
Return mzT from a single file
}
