% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_var}
\alias{order_var}
\title{order_var}
\usage{
order_var(data, g)
}
\arguments{
\item{data}{dataset g g}

\item{g}{g}
}
\description{
Function to order the labels of the socioeconomic and demopraphic variables
}
