% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{order_var_eu}
\alias{order_var_eu}
\title{order_var_eu}
\usage{
order_var_eu(data, g)
}
\arguments{
\item{data}{dataset in which we want to order the labels of the socioeconomic and demographic variables}

\item{g}{variable for which we want to sort the labels}
}
\value{
a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
}
\description{
Function to order the labels of the socioeconomic and demographic variables
}
