% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_vars}
\alias{order_vars}
\title{order_vars}
\usage{
order_vars(data, g)
}
\arguments{
\item{data}{dataset in which we want to order the labels of the socioeconomic and demographic variables}

\item{g}{variable for which we want to sort the labels}
}
\value{
a dataset in which the labels are ordered for the selected socioeconomic or demographic variable
}
\description{
Function to order the labels of the socioeconomic and demographic variables in intersectional graph
}
