% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{MZLOG-OBJ Plot Heat map}
\usage{
plot_heatmap(
  input_mz_obj,
  metadata = NULL,
  annotation = "phenotype",
  title = "HeatMap of intensities",
  save_file = TRUE,
  cluster = NULL,
  ...
)
}
\arguments{
\item{input_mz_obj}{\cr
DataFrame : Input MZ-Obj}

\item{metadata}{\cr
DataFrame : Metadata-Obj of Samples}

\item{annotation}{\cr
String: Colname of metadata to annotate on.}

\item{title}{String: Plot title}

\item{save_file}{String: Path to save the plot.
TRUE: Save to default output directory
NA: Do not save}

\item{cluster}{c("row"=Boolean,"col"=Boolean): Cluster data? Defaults to False}
}
\description{
Create Welch, Fold & Volcano plot, with default inputs
}
