% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_plot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Welch & FoldChange Volcano Plot}
\usage{
plot_volcano(welch, fold_change, title = "Volcano Plot")
}
\arguments{
\item{input_mzlog_obj}{\cr
DataFrame : Log2 of Input MZ-Obj}

\item{phenotype_a}{\cr
DataFrame: Filtered metadata for phenotypeA}

\item{phenotype_b}{\cr
DataFrame: Filtered metadata for phenotypeB}
}
\description{
Compare two phenotypes
}
