% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMRReport.R
\name{plotLikelihoodProfile}
\alias{plotLikelihoodProfile}
\title{Likelihood profile visualization}
\usage{
plotLikelihoodProfile(
  combinedProfile,
  siteProfileList = NULL,
  mrEstimate = NULL,
  title = "Profile Log-Likelihood"
)
}
\arguments{
\item{combinedProfile}{Output of \code{\link{poolLikelihoodProfiles}}.}

\item{siteProfileList}{Optional named list of site profile objects.}

\item{mrEstimate}{Optional output of \code{\link{computeMREstimate}} for
annotating the MLE and CI.}

\item{title}{Plot title. Default is "Profile Log-Likelihood".}
}
\value{
A ggplot2 object.
}
\description{
Creates a ggplot2 visualization of the profile log-likelihood curve(s),
showing individual site profiles (if provided) and the combined profile
with MLE and confidence interval.
}
\details{
Plot Profile Likelihood
}
\examples{
profiles <- simulateSiteProfiles(nSites = 3, trueBeta = 0.5)
combined <- poolLikelihoodProfiles(profiles)
plotLikelihoodProfile(combined, profiles)

}
