% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{price_shock}
\alias{price_shock}
\title{Details: main function to apply a specific price shock to the different COICOP categories}
\usage{
price_shock(data)
}
\description{
Details: main function to apply a specific price shock to the different COICOP categories
}
