% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{rawhbs_eu}
\alias{rawhbs_eu}
\title{rawhbs_eu}
\usage{
rawhbs_eu(year, country = "all", path)
}
\arguments{
\item{year}{year/s of the HBS to process. Available options: 2010, 2015, 2020.}

\item{country}{code of the country or countries of the HBS to process. By default, it processes all available data in the working directory To see the available countries and codes, run country_code().}

\item{path}{Local path to the folder where the HBS's are stored. Not included in the package.}
}
\value{
RData file with the HBS microdata for each selected country and year.
}
\description{
Function to process raw data from the Household Budget Survey (HBS) for each year and country.
}
