% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{rename_coicop}
\alias{rename_coicop}
\title{rename_coicop}
\usage{
rename_coicop(data)
}
\arguments{
\item{data}{dataframe with the hbs data to rename}
}
\value{
data file with the renamed expenditure categories
}
\description{
Function to rename all coicop expenditure categories
}
