% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_values}
\alias{rename_values}
\title{rename_values}
\usage{
rename_values(data, current_var)
}
\arguments{
\item{data}{dataset to be standardized.}

\item{current_var}{column name to be standardized.}
}
\value{
a dataset with labels renamed based in the mapping included in the package.
}
\description{
Function to rename the codified values of the dataset to the meaningful
values detailed in the mapping included in the package.
}
