% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{rename_values_eu}
\alias{rename_values_eu}
\title{rename_values_eu}
\usage{
rename_values_eu(data, map = variables_h, current_var)
}
\arguments{
\item{data}{dataset to be standardized.}

\item{map}{mapping file to be used (included in medusa)}

\item{current_var}{column name to be standardized.}
}
\value{
a dataset with labels renamed based in the mapping included in the package.
}
\description{
Function to rename the codified values of the dataset to the meaningful
values detailed in the mapping included in the package.
}
