% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_mzlog_randomforest.R
\name{rf_select}
\alias{rf_select}
\title{Find Correlation data within a mzLog_obj}
\usage{
rf_select(
  correlation_data,
  metadata,
  attribute = "phenotype",
  feat_size_seq = seq(50, 1000, by = 50)
)
}
\arguments{
\item{correlation_data}{\cr
DataFrame : from rf_correlation}

\item{metadata}{\cr
DataFrame: metadata object}

\item{attribute}{\cr
String: which metadata attribute to compare}

\item{feat_size_seq}{\cr
Sequence to find optimal "number_of_variables"}
}
\value{
caret::rfe object
}
\description{
Find Correlation data within a mzLog_obj
}
