% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Estimate.R
\name{RunReml}
\alias{RunReml}
\title{Run residual maximum likelihood (REML) to fit the linear mixed model.}
\usage{
RunReml(
  sid,
  base,
  bulk,
  rancmp,
  MetaData,
  RanSplit,
  iter_max,
  ct_name,
  bk_name,
  verbose,
  filename = NULL,
  DoParallel
)
}
\arguments{
\item{sid:}{numeric, id of the inputed samples;}

\item{base:}{matrix of the fixed components (gene x cell type);}

\item{bulk:}{matrix of the bulk RNA-seq samples (gene x sample);}

\item{rancmp:}{matrix of the random components (gene x cell);}

\item{MetaData:}{data frame of the summary information of bulk RNA-seq samples. Include sampleID, cellTyp;}

\item{RanSplit:}{character, label for how to seperate random components. Default as \code{NULL}. If \code{NULL}, all cells will be fitted as one random component;}

\item{iter_max:}{numeric, maximum iteration number.Default as 1e+3;}

\item{ct_name:}{vector of cell type names;}

\item{bk_name:}{vector of bulk RNA-seq sample names;}

\item{verbose:}{bool, whether to show running information or not during deconvolution analysis. Default as FALSE;}

\item{filename:}{path, save the warning information during REML analysis;}

\item{DoParallel:}{bool, whether to perform analysis parallelly.}
}
\value{
a list with elements:
  *b: matrix of cell type proportions estimated by mixed linear model (sample x cell type);
  *p: matrix of p value (χ²(df=1)) for the cell type proportions estimated by mixed linear model (sample x cell type);
}
\description{
Run residual maximum likelihood (REML) to fit the linear mixed model.
}
