% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLM_Estimate.R
\name{SCdcv}
\alias{SCdcv}
\title{Single-cell level deconvolution}
\usage{
SCdcv(
  bulk,
  sce,
  select.ct,
  ncpu = NULL,
  smoothing = TRUE,
  gene = NULL,
  op = 0.2,
  maxgene = 1000,
  nbin = 0.1,
  SF = 1000
)
}
\arguments{
\item{bulk}{A matrix containing bulk RNA-Seq data. Each row corresponds to a certain gene and each column to a certain sample.}

\item{sce}{A 'Seurat' object containing the single-cell RNA-Seq data. Meta data of the 'Seurat' object must includes 'cellType' and 'space'.}

\item{select.ct}{A character (or character vector) of the names of the target cell-types. The default value is NULL. With default value, all cell-types in the single-cell data will be used.}

\item{ncpu}{The number of CPU cores to be used.}

\item{smoothing}{A Boolean variable to determine whether to smooth the BLUP along cell-space or not. The default value is TRUE.}

\item{gene}{A character vector of the gene names to use as signature for the deconvolution.The default value is NULL. With default, MLM will select genes to differentiate cells in the target cell-type with ANOVA.}

\item{op}{A numeric variable to determine the overlapping between cell-clusters. The default value is 0.2.}

\item{maxgene}{A numeric variable to determine the maximum number of genes to be selected. The default value is 1e+3.}

\item{nbin}{A numeric variable to determine the number of cell-clusters. The default value is 0.2.}

\item{SF}{Scaling factor. The default value is 1e+3.

#Gene selecting}
}
\value{
A list with cell_density-cell_name matrix:
}
\description{
Single-cell level deconvolution
}
\examples{
library(MLM)
SCdcv(bulk = example.bulk,sce = example.sce,select.ct = 'alpha')
}
