% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shocks}
\alias{shocks}
\title{Example csv file in which the change in prices must be introduced to apply a price shock}
\format{
.csv
}
\source{
medusa
}
\usage{
shocks
}
\description{
Example csv file in which the change in prices must be introduced to apply a price shock
}
\keyword{datasets}
