% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shocks_eu}
\alias{shocks_eu}
\title{Example csv file in which the change in prices must be introduced to apply a price shock to EU microdata}
\format{
.csv
}
\source{
medusa
}
\usage{
shocks_eu
}
\description{
Example csv file in which the change in prices must be introduced to apply a price shock to EU microdata
}
\keyword{datasets}
