% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shocks_eu_s}
\alias{shocks_eu_s}
\title{Example csv file in which the change in prices must be introduced to apply multiple scenarios and price shocks to EU microdata}
\format{
.csv
}
\source{
medusa
}
\usage{
shocks_eu_s
}
\description{
Example csv file in which the change in prices must be introduced to apply multiple scenarios and price shocks to EU microdata
}
\keyword{datasets}
