% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{standardize_eu}
\alias{standardize_eu}
\title{standardize_eu}
\usage{
standardize_eu(data, map = variables_h)
}
\arguments{
\item{data}{dataset to be standardized.}

\item{map}{mapping file to be used (included in medusa)}
}
\value{
a dataset with the variables and labels renamed based in the mapping included in the package.
}
\description{
Function to standarize data names.
}
