% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_academicapplied.R
\name{step_academicapplied}
\alias{step_academicapplied}
\title{Categorize whether a member works in an applied or academic setting.}
\usage{
step_academicapplied(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble
}
\description{
Categorize whether a member works in an applied or academic setting.
}
\examples{
#demo_raw <- step_academicapplied(df = demo_raw)
}
