% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_agegroup.R
\name{step_agegroup}
\alias{step_agegroup}
\title{Clean and categorize a member's age. Bins have been pre-established according
to best practices. Age is calculated from the date the function was ran.
Results may therefore vary.}
\usage{
step_agegroup(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
Age, AgeGroup
}
\description{
Clean and categorize a member's age. Bins have been pre-established according
to best practices. Age is calculated from the date the function was ran.
Results may therefore vary.
}
\examples{
#step_agegroup(df = demo_raw)
}
