% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_birthdate.R
\name{step_birthdate}
\alias{step_birthdate}
\title{Re-factor the Birthdate column to be a true Date variable}
\usage{
step_birthdate(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble
}
\description{
Re-factor the Birthdate column to be a true Date variable
}
\examples{
#demo_raw <- step_birthdate(df = demo_raw)
}
