% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_city.R
\name{step_city}
\alias{step_city}
\title{Cleans and categorizes open-ended city name data and corrects errors.}
\usage{
step_city(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble
}
\description{
Cleans and categorizes open-ended city name data and corrects errors.
}
\examples{
#demo_raw <- step_city(df = demo_raw)
}
