% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_country.R
\name{step_country}
\alias{step_country}
\title{Cleans a member's country of residence/location.}
\usage{
step_country(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column Country
}
\description{
Cleans a member's country of residence/location.
}
\examples{
#demo_raw <- step_country(df = demo_raw)
}
