% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_dues.R
\name{step_dues}
\alias{step_dues}
\title{Cleans a variety of columns in dues data}
\usage{
step_dues(df, year)
}
\arguments{
\item{df}{a tibble of dues data}

\item{year}{the SIOP year to filter data on}
}
\value{
a tibble of SIDs applicable for the given year selected.
}
\description{
Cleans a variety of columns in dues data
}
\examples{
#dues_cleaned <- step_dues(df = dues_cleaned, year = 2020)
}
