% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ethnicity.R
\name{step_ethnicity}
\alias{step_ethnicity}
\title{Title}
\usage{
step_ethnicity(df, detailed_types = FALSE)
}
\arguments{
\item{df}{a raw demographic tibble}

\item{detailed_types}{If ethnicity should be reported by country, or according
to the new simplified schema such as Black, Hispanic, White, etc.}
}
\value{
tibble
}
\description{
Title
}
