% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_field.R
\name{step_field}
\alias{step_field}
\title{Cleans a member's field of study}
\usage{
step_field(df, detailed_types = FALSE)
}
\arguments{
\item{df}{a tibble of demographic data}

\item{detailed_types}{logical, whether to report country-based ethnicity or not}
}
\value{
a tibble with cleaned column Field
}
\description{
Cleans a member's field of study
}
\examples{
#demo_raw <- step_field(df = demo_raw, detailed_types = FALSE)
}
