% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_gender.R
\name{step_gender}
\alias{step_gender}
\title{Cleans a member's gender}
\usage{
step_gender(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column Gender
}
\description{
Cleans a member's gender
}
\examples{
#demo_raw <- step_gender(df = demo_raw)
}
