% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_interests.R
\name{step_interests}
\alias{step_interests}
\title{Function that cleans the primary interests fields of the original
demographics dataset, up to a total of 3 unique interest fields.}
\usage{
step_interests(df)
}
\arguments{
\item{df}{Specify the raw xlsx demographics file}
}
\value{
tibble
}
\description{
Function that cleans the primary interests fields of the original
demographics dataset, up to a total of 3 unique interest fields.
}
\examples{
#step_interests(df)
}
