% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_involvement.R
\name{step_involvement}
\alias{step_involvement}
\title{Cleans Involvement Sought field}
\usage{
step_involvement(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
tibble
}
\description{
Cleans Involvement Sought field
}
\examples{
#step_involvement(df = demo_raw)
}
