% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_membergroup.R
\name{step_membergroup}
\alias{step_membergroup}
\title{Cleans a member's categorization as individual or prospect}
\usage{
step_membergroup(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column MemberGroup
}
\description{
Cleans a member's categorization as individual or prospect
}
\examples{
#demo_raw <- step_membergroup(df = demo_raw)
}
