% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_other_orgs.R
\name{step_other_orgs}
\alias{step_other_orgs}
\title{Cleans Other Organizational Affiliations and licensure fields}
\usage{
step_other_orgs(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
tibble
}
\description{
Cleans Other Organizational Affiliations and licensure fields
}
\examples{
#step_other_orgs(df = demo_raw)
}
