% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_privatepublic.R
\name{step_privatepublic}
\alias{step_privatepublic}
\title{Cleans a member's school categorization}
\usage{
step_privatepublic(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column Priv_Publ
}
\description{
Cleans a member's school categorization
}
\examples{
#demo_raw <- step_privatepublic(df = demo_raw)
}
