% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_profitnon.R
\name{step_profitnon}
\alias{step_profitnon}
\title{Cleans a member's school categorization.}
\usage{
step_profitnon(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column Profit_Non
}
\description{
Cleans a member's school categorization.
}
\examples{
#demo_raw <- step_profitnon(df = demo_raw)
}
