% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sid.R
\name{step_sid}
\alias{step_sid}
\title{Removes SID duplicates}
\usage{
step_sid(df)
}
\arguments{
\item{df}{a tibble of dues data}
}
\value{
a tibble of rows with unique SID values
}
\description{
Removes SID duplicates
}
