% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_state.R
\name{step_state}
\alias{step_state}
\title{Cleans a member's state of residence, USA only.}
\usage{
step_state(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column State_US
}
\description{
Cleans a member's state of residence, USA only.
}
\examples{
#demo_raw <- step_state(df = demo_raw)
}
