% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_eu.R
\name{update_year}
\alias{update_year}
\title{update_year}
\usage{
update_year(data, new_year)
}
\arguments{
\item{data}{dataframe with the hbs data to update (The COICOP codes in the dataframe must be renamed beforehand using rename_coicop.)}

\item{new_year}{year to which you want to update the microdata}
}
\value{
data file with the updated expenditure data
}
\description{
Function to update the expenditure data of the HBS
}
