% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validateSiteProfile}
\alias{validateSiteProfile}
\title{Validate Site Profile Object}
\usage{
validateSiteProfile(siteProfile)
}
\arguments{
\item{siteProfile}{A list to validate.}
}
\value{
Invisible TRUE if valid; throws an error otherwise.
}
\description{
Checks that a site profile object has the required structure.
}
\keyword{internal}
