% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_m_2010}
\alias{variables_m_2010}
\title{Mapping for 2010 member file}
\format{
.csv
}
\source{
medusa
}
\usage{
variables_m_2010
}
\description{
Mapping for 2010 member file
}
\keyword{datasets}
