% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_m_2015}
\alias{variables_m_2015}
\title{Mapping for 2015 member file}
\format{
.csv
}
\source{
medusa
}
\usage{
variables_m_2015
}
\description{
Mapping for 2015 member file
}
\keyword{datasets}
