% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted.median}
\alias{weighted.median}
\title{weighted.median}
\usage{
weighted.median(x, w, na.rm = TRUE, type = 2, collapse = TRUE)
}
\arguments{
\item{x}{Data values. A vector pf numeric values, for which the median or quantiles are required.}

\item{w}{Weights. A vector of nonnegative numbers, of the same lenght as x.}

\item{na.rm}{Logical. Whether to ignore NA values.}

\item{type}{Integer specifying the rule for calculating the median or quantile, corresponding to the rules available for quantile. The only valid choices are type=1, 2 or 4.}

\item{collapse}{Research use only.}
}
\description{
Function to calculate weighted median. Extracted from package spatstat.geom v3.2-5
}
