% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted.quantile}
\alias{weighted.quantile}
\title{weighted.quantile}
\usage{
weighted.quantile(
  x,
  w,
  probs = seq(0, 1, 0.25),
  na.rm = TRUE,
  type = 4,
  collapse = TRUE
)
}
\arguments{
\item{x}{Data values. A vector pf numeric values, for which the median or quantiles are required.}

\item{w}{Weights. A vector of nonnegative numbers, of the same lenght as x.}

\item{probs}{Probabilities for which the quantiles should be computed. A numeric vector of values between 0 and 1.}

\item{na.rm}{Logical. Whether to ignore NA values.}

\item{type}{Integer specifying the rule for calculating the median or quantile, corresponding to the rules available for quantile. The only valid choices are type=1, 2 or 4.}

\item{collapse}{Research use only.}
}
\description{
Function to calculate weighted quantile. Extracted from package spatstat.geom v3.2-5
}
