% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_var_impact}
\alias{available_var_impact}
\title{Details: main function to calculate distributional impacts for different price shocks}
\usage{
available_var_impact()
}
\description{
Details: main function to calculate distributional impacts for different price shocks
}
