% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{available_var_intersec}
\alias{available_var_intersec}
\title{Details: function that returns the variables for which the calculation of intersectional distributional impacts is available}
\usage{
available_var_intersec()
}
\description{
Details: function that returns the variables for which the calculation of intersectional distributional impacts is available
}
