% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elevate_hbs}
\alias{elevate_hbs}
\title{Details: main function to elevate the Spanish Household Budget Survey (HBS)}
\usage{
elevate_hbs(data, year, country = "ES")
}
\arguments{
\item{data}{description}

\item{year}{year of the HBS you want to elevate}
}
\description{
Details: main function to elevate the Spanish Household Budget Survey (HBS)
}
