% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epf_2021_m}
\alias{epf_2021_m}
\title{Members file of the 2021 Household Budget Survey (INE), which collects detailed sociodemographic information on the members of each household.}
\format{
.csv
}
\source{
medusa
}
\usage{
epf_2021_m
}
\description{
Members file of the 2021 Household Budget Survey (INE), which collects detailed sociodemographic information on the members of each household.
}
\keyword{datasets}
