% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conf_attendance.R
\name{get_conf_attendance}
\alias{get_conf_attendance}
\title{Cleans a tibble according to who attended a SIOP conference}
\usage{
get_conf_attendance(df, merge_demo_from, year)
}
\arguments{
\item{df}{SIOP conference data frame}

\item{merge_demo_from}{Typically the tibble obtained from make_final_data()}

\item{year}{Which conferene year to analyze.}
}
\value{
a tibble of conference attendees
}
\description{
Cleans a tibble according to who attended a SIOP conference
}
