% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impact_intersectional}
\alias{impact_intersectional}
\title{Details: main function to calculate the distributional impacts based in one or more socioeconomic or demographic variables}
\usage{
impact_intersectional(data, pairs = is_categories)
}
\arguments{
\item{data}{input data calculate}

\item{var1}{first socioeconomic or demographic variable for which intersectional distributional impact is calculated. By default: all variables of category_a in is_categories.}

\item{var2}{second socioeconomic or demographic variable for which intersectional distributional impact is calculated. By default: all variables of category_b in is_categories.}
}
\description{
Details: main function to calculate the distributional impacts based in one or more socioeconomic or demographic variables
}
