% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_rawhbs}
\alias{load_rawhbs}
\title{Details: main function to load the Spanish Household Budget Survey (HBS)}
\usage{
load_rawhbs(year)
}
\arguments{
\item{year}{year of the HBS you want to load}
}
\value{
a list with the 3 files of the HBS
}
\description{
Details: main function to load the Spanish Household Budget Survey (HBS)
}
