% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medusa.R
\docType{package}
\name{medusa}
\alias{medusa-package}
\alias{medusa}
\title{medusa}
\description{
\itemize{
\item Github: https://github.com/bc3LC/medusa
\item Webpage: https://bc3lc.github.io/medusa/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bc3LC/medusa}
  \item Report bugs at \url{https://github.com/bc3LC/medusa/issues}
}

}
\author{
\strong{Maintainer}: Eva Alonso-Epelde \email{eva.alonso@bc3research.org} (\href{https://orcid.org/0000-0002-4052-1703}{ORCID})

Authors:
\itemize{
  \item Clàudia Rodés-Bachs \email{claudia.rodes@bc3research.org} (\href{https://orcid.org/0000-0001-6696-7685}{ORCID})
  \item María Moyano-Reina \email{maria.moyano@bc3research.org } (\href{https://orcid.org/0009-0000-4290-0697}{ORCID})
  \item Jon Sampedro \email{jon.sampedro@bc3research.org} (\href{https://orcid.org/0000-0002-2277-1530}{ORCID})
  \item Xaquín García-Muros \email{xaquin.garcia@bc3research.org } (\href{https://orcid.org/0000-0002-2667-3292}{ORCID})
}

}
