% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shocks}
\alias{shocks}
\title{List to apply a specific price shock to the different COICOP categories}
\format{
.csv
}
\source{
medusa
}
\usage{
shocks
}
\description{
List to apply a specific price shock to the different COICOP categories
}
\keyword{datasets}
