% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{standardize}
\usage{
standardize(data)
}
\arguments{
\item{data}{dataset to be standardized.}
}
\value{
a dataset with the variables and labels renamed based in the mapping included in the package.
}
\description{
Function to standarize data names.
}
