% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ethnicity.R
\name{step_ethnicity}
\alias{step_ethnicity}
\title{Cleans a member's racial ethnicity.}
\usage{
step_ethnicity(df, detailed_types = FALSE)
}
\arguments{
\item{df}{a tibble of demographic data}

\item{detailed_types}{If ethnicity should be reported by country,
or according to the new simplified schema such as Black, Hispanic, White, etc.}
}
\value{
a tibble with cleaned column Ethnicity
}
\description{
Cleans a member's racial ethnicity.
}
\examples{
#demo_raw <- step_ethnicity(df = demo_raw, detailed_types = FALSE)
}
