% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_highestdegree.R
\name{step_highestdegree}
\alias{step_highestdegree}
\title{Cleans a member's highest earned degree.}
\usage{
step_highestdegree(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with cleaned column Highest_Degree
}
\description{
Cleans a member's highest earned degree.
}
