% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_interests.R
\name{step_interests}
\alias{step_interests}
\title{Cleans primary interests categories, up to 3 in total}
\usage{
step_interests(df)
}
\arguments{
\item{df}{Specify the raw xlsx demographics file}
}
\value{
tibble
}
\description{
Cleans primary interests categories, up to 3 in total
}
\examples{
#step_interests(df)
}
