% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_involvement.R
\name{step_involvement}
\alias{step_involvement}
\title{Function that cleans the Involvement Sought field of the original
demographics dataset.}
\usage{
step_involvement(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
tibble
}
\description{
Function that cleans the Involvement Sought field of the original
demographics dataset.
}
\examples{
#step_involvement(df = demo_raw)
}
