% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_other_orgs.R
\name{step_other_orgs}
\alias{step_other_orgs}
\title{Function that cleans the Other Organizational Affiliations field of the original
demographics dataset.}
\usage{
step_other_orgs(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
tibble
}
\description{
Function that cleans the Other Organizational Affiliations field of the original
demographics dataset.
}
\examples{
#step_other_orgs(df = demo_raw)
}
