% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_sector.R
\name{step_sector}
\alias{step_sector}
\title{Cleans a member's primary employment sector.}
\usage{
step_sector(df)
}
\arguments{
\item{df}{a tibble of demographic data}
}
\value{
a tibble with two cleaned columns: "Primary_Occupation" which is
a more thorough breakdown of a member's career, and "Primary_Sector", which
is categorized into four main buckets: Academy, Government, Private, Other
}
\description{
Cleans a member's primary employment sector.
}
\examples{
#demo_raw <- step_sector(df = demo_raw)
}
